+++ src/get_attribute.ts
@@ -1,8 +1,9 @@
 import { maybeOf, Maybe } from '@fluss/core';
+import { AttributeNamesOf, GlobalAttributeNames } from './types';
 
-export function getAttribute(
-  element: Element,
-  name: string
+export function getAttribute<E extends Element>(
+  element: E,
+  name: AttributeNamesOf<E> | GlobalAttributeNames
 ): Maybe<string> {
   return maybeOf(element.getAttribute(name));
 }