+++ src/has_attribute.ts
@@ -1,3 +1,8 @@
-export function hasAttribute(element: Element, name: string): boolean {
+import { AttributeNamesOf, GlobalAttributeNames } from './types';
+
+export function hasAttribute<E extends Element>(
+  element: E,
+  name: AttributeNamesOf<E> | GlobalAttributeNames
+): boolean {
   return element.hasAttribute(name);
 }