+++ src/set_attribute.ts
@@ -1,6 +1,8 @@
-export function setAttribute(
-  element: Element,
-  key: string,
+import { AttributeNamesOf, GlobalAttributeNames } from './types';
+
+export function setAttribute<E extends Element>(
+  element: E,
+  key: AttributeNamesOf<E> | GlobalAttributeNames,
   value: string
 ): void {
   element.setAttribute(key, value);