+++ src/toggle_attribute.ts
@@ -1,6 +1,8 @@
-export function toggleAttribute(
-  element: Element,
-  name: string,
+import { GlobalAttributeNames, AttributeNamesOf } from './types';
+
+export function toggleAttribute<E extends Element>(
+  element: E,
+  name: AttributeNamesOf<E> | GlobalAttributeNames,
   force?: boolean
 ): boolean {
   return element.toggleAttribute(name, force);